clean code + refactor
