+++ Source/Controllers/Auth/RefreshToken.cs
@@ -46,13 +46,13 @@ public class RefreshTokenController : ControllerBase
             var responseBody = response.Content.ReadAsStringAsync().Result;
             Response.ContentType = "application/json";
 
-            JObject responseJson = JObject.Parse(responseBody);
-            JObject resultJson = new JObject();
+            var responseJson = JObject.Parse(responseBody);
+            var resultJson = new JObject();
             resultJson["access_token"] = responseJson["id_token"];
             resultJson["refresh_token"] = responseJson["refresh_token"];
             resultJson["expires_in"] = responseJson["expires_in"];
 
-            return new ObjectResult(resultJson); 
+            return new ObjectResult(resultJson);
         }
         catch (Exception ex)
         {
