+++ Source/Controllers/Rooms/SearchRooms.cs
@@ -3,8 +3,9 @@
 using System.Net;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.ModelBinding;
-using PoliFemoBackend.Source.Utils;
 using Newtonsoft.Json.Linq;
+using PoliFemoBackend.Source.Utils;
+
 #endregion
 
 namespace PoliFemoBackend.Source.Controllers.Rooms;
@@ -14,7 +15,6 @@ namespace PoliFemoBackend.Source.Controllers.Rooms;
 [ApiExplorerSettings(GroupName = "Rooms")]
 [Route("v{version:apiVersion}/rooms/search")]
 [Route("/rooms/search")]
-
 public class SearchRoomsController : ControllerBase
 {
     /// <summary>
@@ -42,20 +42,24 @@ public class SearchRoomsController : ControllerBase
                 StatusCode = (int)HttpStatusCode.InternalServerError
             };
         }
+
         var t4 = RoomUtil.GetFreeRooms(t3[0], hourStart, hourStop);
         if (t4 is null || t4.Count == 0) return NoContent();
         var results = new JArray();
-        foreach(var room in t4)
-            if(room != null){
-                JObject formattedRoom = JObject.FromObject(room);
+        foreach (var room in t4)
+            if (room != null)
+            {
+                var formattedRoom = JObject.FromObject(room);
                 var roomLink = formattedRoom.GetValue("link");
-                if(roomLink != null){
-                    int roomId = Int32.Parse(roomLink.ToString().Split("idaula=")[1]);
+                if (roomLink != null)
+                {
+                    var roomId = int.Parse(roomLink.ToString().Split("idaula=")[1]);
                     formattedRoom.Add(new JProperty("room_id", roomId));
                 }
+
                 results.Add(formattedRoom);
             }
-            
+
         var json = new { freeRooms = results };
         return Ok(new JObject(new JProperty("freeRooms", results)));
     }
