+++ Source/Data/Constants.cs
@@ -19,7 +19,8 @@ public static class Constants
 
     public const string AzureOrgIssuer = "https://login.microsoftonline.com/0a17712b-6df3-425d-808e-309df28a5eeb/v2.0";
 
-    public const string AzureCommonIssuer = "https://login.microsoftonline.com/9188040d-6c67-4c5b-b112-36a304b66dad/v2.0";
+    public const string AzureCommonIssuer =
+        "https://login.microsoftonline.com/9188040d-6c67-4c5b-b112-36a304b66dad/v2.0";
 
     public const string Authorization = "Authorization";
     public const double MaxRate = 5;
