+++ Source/Utils/RoomUtil.cs
